+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/TaskInstaller.java
@@ -4,6 +4,7 @@ import io.dropwizard.servlets.tasks.Task;
 import io.dropwizard.setup.Environment;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.install.InstanceInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.order.Order;
 import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
 
 /**
@@ -13,6 +14,7 @@ import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
  * @author Vyacheslav Rusakov
  * @since 01.09.2014
  */
+@Order(70)
 public class TaskInstaller implements FeatureInstaller<Task>, InstanceInstaller<Task> {
 
     @Override